同步上游

+++ NsisoLauncher/NsisoLauncherCore/Net/MultiThreadDownloader.cs
@@ -75,6 +75,7 @@ namespace NsisoLauncherCore.Net
         public int ProcessorSize { get; set; } = 5;
         public bool IsBusy { get; private set; } = false;
         public WebProxy Proxy { get; set; }
+        public bool CheckFileHash { get; set; }
 
         public event EventHandler<DownloadProgressChangedArg> DownloadProgressChanged;
         public event EventHandler<DownloadSpeedChangedArg> DownloadSpeedChanged;
@@ -153,6 +154,7 @@ namespace NsisoLauncherCore.Net
                 {
                     cancellationTokenSource = new CancellationTokenSource();
                     IsBusy = true;
+                    _errorList.Clear();
                     if (ProcessorSize == 0)
                     {
                         IsBusy = false;
@@ -233,12 +235,13 @@ namespace NsisoLauncherCore.Net
                         ApendDebugLog("下载完成:" + item.From);
 
                         #region 校验
-                        if (item.Checker != null)
+                        if (CheckFileHash && item.Checker != null)
                         {
                             item.SetState("校验中");
                             if (!item.Checker.CheckFilePass())
                             {
-                                ApendDebugLog("校验失败");
+                                item.SetState("校验失败");
+                                ApendDebugLog(string.Format("{0}校验哈希值失败，目标哈希值:{1}", item.TaskName, item.Checker.CheckSum));
                                 File.Delete(item.To);
                                 if (!_errorList.ContainsKey(item))
                                 {
@@ -248,6 +251,7 @@ namespace NsisoLauncherCore.Net
                             else
                             {
                                 item.SetState("校验成功");
+                                ApendDebugLog(string.Format("{0}校验哈希值成功:{1}", item.TaskName, item.Checker.CheckSum));
                             }
                         }
                         #endregion
@@ -315,27 +319,42 @@ namespace NsisoLauncherCore.Net
                 Stream responseStream = response.GetResponseStream();
                 responseStream.ReadTimeout = 5000;
                 FileStream fs = new FileStream(buffFilename, FileMode.Create);
-                byte[] bArr = new byte[1024];
-                int size = responseStream.Read(bArr, 0, (int)bArr.Length);
 
-                while (size > 0)
+                try
                 {
-                    if (cancelToken.IsCancellationRequested)
+                    byte[] bArr = new byte[1024];
+                    int size = responseStream.Read(bArr, 0, (int)bArr.Length);
+
+                    while (size > 0)
                     {
-                        ApendDebugLog("放弃下载:" + task.TaskName);
-                        fs.Close();
-                        responseStream.Close();
-                        RemoveItemFromViewTask(task);
-                        return;
+                        if (cancelToken.IsCancellationRequested)
+                        {
+                            ApendDebugLog("放弃下载:" + task.TaskName);
+                            fs.Close();
+                            responseStream.Close();
+                            RemoveItemFromViewTask(task);
+                            return;
+                        }
+                        fs.Write(bArr, 0, size);
+                        size = responseStream.Read(bArr, 0, (int)bArr.Length);
+                        _downloadSizePerSec += size;
+                        task.IncreaseDownloadSize(size);
                     }
-                    fs.Write(bArr, 0, size);
-                    size = responseStream.Read(bArr, 0, (int)bArr.Length);
-                    _downloadSizePerSec += size;
-                    task.IncreaseDownloadSize(size);
+
+                    fs.Close();
+                    responseStream.Close();
+                    File.Move(buffFilename, realFilename);
                 }
-                fs.Close();
-                responseStream.Close();
-                File.Move(buffFilename, realFilename);
+                catch (Exception e)
+                {
+                    fs.Close();
+                    responseStream.Close();
+                    ApendErrorLog(e);
+                    if (!_errorList.ContainsKey(task))
+                    {
+                        _errorList.Add(task, e);
+                    }
+                }
             }
             catch (Exception e)
             {
